// src/app/shared/interfaces/resumo-avarias.model.ts
export interface ResumoAvariasMV {
  dataAbertura?: number;
  ordemServico?: any;
  evento?: any;
  percentualProtecaoAvaria?: any;
  requisicaoId?: number;
  placa?: string;
  modelo?: string;
  contratoMasterId?: number;
  nomeFantasia?: string;
  telefone?: string;
  dataAtendimento?: any;
  valorFipe?: number;
  km?: number;
  franquia?: any;
  prazoLiberacao?: any;
  prazoLiberacaoExtenso?: any;
  valorCobrado?: any;
  possiveisCausasProblemas?: any;
  impactoNaoExecucaoTrocaPeca?: any;
  atendimentoFornecedorId?: number;
}